Refactored to fix circular imports
